requirements.txt
Django==1.9.8
django-autofixture==0.12.1
django-autoslug==1.9.3
django-extensions==1.6.7
Pillow==3.3.0
six==1.10.0
Werkzeug==0.11.10
